# api-doc.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: VinDr Lab API Document
  contact:
    name: VinDr Lab Development Team
  license:
    name: MIT
paths:
  /labels:
    get:
      description: get Labels by array
      operationId: fetchLabels
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: label_group_id
          in: query
          required: false
          style: form
          schema:
            type: array
            items:
              type: string
              format: uuid
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          impression:
                            type: array
                            items:
                              $ref: "#/components/schemas/Label"
                          finding:
                            type: array
                            items:
                              $ref: "#/components/schemas/Label"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: |
        create a new Label

        id of object can be empty
      operationId: createLabel
      parameters:
        - $ref: "#/components/parameters/authParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Label"
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          label_id:
                            type: string
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /labels/{label_id}:
    put:
      description: update a Label by its id
      operationId: updateLabel
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: label_id
          in: path
          description: ID of Label to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Label"
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          label_id:
                            type: string
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      description: delete a Label by its id
      operationId: deleteLabel
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: label_id
          in: path
          description: ID of Label to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          label_id:
                            type: string
                            format: uuid
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /label_groups:
    get:
      description: get LabelGroups by array
      operationId: fetchLabelGroups
      parameters:
        - $ref: "#/components/parameters/authParam"
      responses:
        "200":
          description: get LabelGroups response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/LabelGroup"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: |
        create a new LabelGroup

        id of object can be empty
      operationId: createLabelGroup
      parameters:
        - $ref: "#/components/parameters/authParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelGroup"
      responses:
        "200":
          description: get LabelGroup response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /label_groups/{label_group_id}:
    put:
      description: update LabelGroup
      operationId: updateLabelGroup
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: label_group_id
          schema:
            type: string
          in: path
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelGroup"
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      description: delete a LabelGroup
      operationId: deleteLabelGroup
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: label_group_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: get labels response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /label_groups/{label_group_id}/labels:
    post:
      parameters:
        - $ref: "#/components/parameters/authParam"
      description: import labels to label group
      operationId: importLabels
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      parameters:
        - $ref: "#/components/parameters/authParam"
      description: export labels from label group
      operationId: exportLabels
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    authParam:
      name: Authorization
      in: header
      schema:
        type: string
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          enum:
            - value: 0
              name: Success
            - value: -1
              name: Failed
            - value: -2
              name: Invalid Data
            - value: -3
              name: Server Error
        message:
          type: string
        count:
          type: integer
          format: int32
          description: size of returned data
...
    Label:
      type: object
      required:
        - name
        - scope
        - color
        - label_group_id
      properties:
        id:
          type: string
          format: uuid
        parent_label_id:
          type: string
          format: uuid
        name:
          type: string
        tag:
          type: string
        type:
          type: string
          enum: [IMPRESSION, FINDING]
        scope:
          type: string
          enum: [STUDY, SERIES, IMAGE]
        short_name:
          type: string
        creator_id:
          type: string
          format: uuid
        description:
          type: string
        color:
          type: string
        annotation_type:
          type: string
          enum: [BOUNDING_BOX, MASK, POLYGON, TAG]
        label_group_id:
          type: string
          format: uuid
...
    LabelGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        color:
          type: string
        name:
          type: string
        creator_id:
          type: string
          format: uuid
        owner_ids:
          type: array
          items:
            type: string
            format: uuid
...